fix finding init files when not in $HOME
I tried to fix the portability problem between linux and macosx in
readlink by not fully expanding the link. This worked find when starting
the shell (since your were in $HOME and the relative path work), but
when you were not n $HOME it failed.

We fix this by now relying on greadlink on macosx.